Move App into components directory and bookform into containers directory
