+++ src/index.js
@@ -5,7 +5,7 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
-import App from './App';
+import App from './components/App';
 import reportWebVitals from './reportWebVitals';
 import reducers from './reducers';
